Refactoring and added initial sheets wrapper class
