+++ requirements.txt
@@ -1,16 +1,29 @@
 aiohttp==3.5.4
 async-timeout==3.0.1
 attrs==19.1.0
+cachetools==3.1.0
 certifi==2019.3.9
 chardet==3.0.4
 discord.py==1.0.1
+google-api-python-client==1.7.8
+google-auth==1.6.3
+google-auth-httplib2==0.0.3
+google-auth-oauthlib==0.3.0
+httplib2==0.12.3
 idna==2.8
 idna-ssl==1.1.0
 multidict==4.5.2
+oauthlib==3.0.1
+pyasn1==0.4.5
+pyasn1-modules==0.2.5
 requests==2.21.0
+requests-oauthlib==1.2.0
+rsa==4.0
+six==1.12.0
 tabulate==0.8.3
 tbapy==1.3.1
 typing-extensions==3.7.2
+uritemplate==3.0.0
 urllib3==1.24.3
 websockets==6.0
 yarl==1.3.0
